Refactor chart types
