+++ app/components/charts/type-charts/base-chart-type.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/charts/type-charts/base-chart-type';
